Backend/app/requirements.txt
Flask
pymongo
Marshmallow
python-dotenv
flask-jwt-extended
werkzeug